backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
diff-match-patch==20241021
dj-rest-auth==7.0.1
Django==5.1.4
django-allauth==65.3.1
django-cors-headers==4.6.0
django-filter==24.3
django-import-export==4.3.4
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
idna==3.10
jwt==1.3.1
Markdown==3.7
packaging==24.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
tablib==3.7.0
urllib3==2.3.0
